+++ polaris-synchronizer/api/src/main/java/org/apache/polaris/tools/sync/polaris/catalog/ETagManager.java
@@ -20,17 +20,27 @@ package org.apache.polaris.tools.sync.polaris.catalog;
 
 import org.apache.iceberg.catalog.TableIdentifier;
 
-public class NotModifiedException extends RuntimeException {
+/**
+ * Generic interface to provide and store ETags for tables within catalogs. This allows the storage
+ * of the ETag to be completely independent from the tool.
+ */
+public interface ETagManager {
 
-  public NotModifiedException(TableIdentifier tableIdentifier) {
-    super("Table " + tableIdentifier + " was not modified.");
-  }
+  /**
+   * Retrieves the ETag for the table.
+   *
+   * @param catalogName the catalog the table is in
+   * @param tableIdentifier the table identifier
+   * @return The ETag for the last known metadata for the table
+   */
+  String getETag(String catalogName, TableIdentifier tableIdentifier);
 
-  public NotModifiedException(String message) {
-    super(message);
-  }
-
-  public NotModifiedException(String message, Throwable cause) {
-    super(message, cause);
-  }
+  /**
+   * After table loading, stores the fetched ETag.
+   *
+   * @param catalogName the catalog the table is in
+   * @param tableIdentifier the table identifier
+   * @param etag the ETag that was provided by the Iceberg REST api
+   */
+  void storeETag(String catalogName, TableIdentifier tableIdentifier, String etag);
 }
