+++ polaris-synchronizer/cli/src/main/java/org/apache/polaris/tools/sync/polaris/options/SourcePolarisOptions.java
@@ -20,6 +20,10 @@ package org.apache.polaris.tools.sync.polaris.options;
 
 import picocli.CommandLine;
 
+/**
+ * Prefixes service_admin connection option names with "source" tags to identify that these are
+ * the connection properties for the source instance.
+ */
 public class SourcePolarisOptions extends BasePolarisOptions {
 
   @Override
